libdmabufinfo: Make DMA buf size uint64_t
DMA buf sysfs stats currently uses unsigned int for the buffer
size. This means it will only be able to account correctly buffers
of size 4GB or less. Change this to uint64_t.

Test: dambuf_dump
Bug: 234753494

+++ libdmabufinfo/dmabufinfo.cpp
@@ -241,7 +241,7 @@ bool ReadDmaBufMapRefs(pid_t pid, std::vector<DmaBuffer>* dmabufs,
         // buffer size.
         //
         // Attempt to retrieve the real buffer size from sysfs.
-        unsigned int size = 0;
+        uint64_t size = 0;
         if (!sysfs_stats || !ReadBufferSize(mapinfo.inode, &size, dmabuf_sysfs_path)) {
             size = mapinfo.end - mapinfo.start;
         }
